// src/enums/step.ts
export const enum Step {
  INVALID = 0,
  BEGIN_FIRST = 1,
  BEGIN_SHUFFLE = 2,
  BEGIN_DRAW = 3,
  BEGIN_MULLIGAN = 4,
  MAIN_BEGIN = 5,
  MAIN_READY = 6,
  MAIN_START_TRIGGERS = 7,
  MAIN_RESOURCE = 8,
  MAIN_DRAW = 9,
  MAIN_START = 10,
  MAIN_ACTION = 11,
  MAIN_COMBAT = 12,
  MAIN_END = 13,
  MAIN_NEXT = 14,
  FINAL_WRAPUP = 15,
  FINAL_GAMEOVER = 16,
  MAIN_CLEANUP = 17,
  MAIN_START_MULLIGAN = 18,
}